requirements.txt
appnope==0.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.3.9
chardet==3.0.4
decorator==4.4.0
Django==2.2
django-extensions==2.1.6
docutils==0.14
idna==2.8
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
pkginfo==1.5.0.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pyaes==1.6.1
pycrypto==2.6.1
Pygments==2.3.1
pytz==2019.1
PyYAML==5.1
readme-renderer==24.0
requests==2.21.0
requests-toolbelt==0.9.1
six==1.12.0
sqlparse==0.3.0
tqdm==4.31.1
traitlets==4.3.2
twine==1.13.0
urllib3==1.24.2
wcwidth==0.1.7
webencodings==0.5.1
